pyproject.toml
[project]
name = "docketbird-mcp"  
version = "0.1.0"
dependencies = [
    "mcp>=1.6.0",
    "requests>=2.31.0", 
    "python-dotenv>=1.0.0",
    "fastapi",
    "uvicorn",
    "pydantic",
    "termcolor",
    "starlette>=0.21.0",
    "pydantic-ai",
    "logfire",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
